feat(dict): update dictionary
This uses the EWDSLIST.GEN file included in the original wordsall.zip
download, rather than using DICTLINE.GEN to create EWDSLIST.GEN. (It
seems like the DICTLINE.GEN is missing a lot of English words that are
present in EWDSLIST.GEN).

The practical effect is that this increases the number of English words
by about 10x, and should bring parity with other Words-based software.